requirements.txt
Adafruit_MotorHAT
click
pyserial
pytest
pyyaml
RPi.GPIO
